+++ modules/Package.swift
@@ -60,7 +60,7 @@ let package = Package(
         .library(name: "ZcashSDKEnvironment", targets: ["ZcashSDKEnvironment"])
     ],
     dependencies: [
-        .package(url: "https://github.com/pointfreeco/swift-composable-architecture", from: "1.4.0"),
+        .package(url: "https://github.com/pointfreeco/swift-composable-architecture", from: "1.4.2"),
         .package(url: "https://github.com/pointfreeco/swift-case-paths", from: "1.1.0"),
         .package(url: "https://github.com/pointfreeco/swift-url-routing", from: "0.6.0"),
         .package(url: "https://github.com/zcash-hackworks/MnemonicSwift", from: "2.2.4"),
@@ -97,6 +97,7 @@ let package = Package(
         .target(
             name: "BalanceBreakdown",
             dependencies: [
+                "BalanceFormatter",
                 "Generated",
                 "DerivationTool",
                 "MnemonicClient",