+++ secantTests/BalanceBreakdownTests/BalanceBreakdownTests.swift
@@ -12,10 +12,12 @@ import Combine
 import Utils
 import Generated
 import BalanceBreakdown
+import Models
 @testable import secant_testnet
 
+@MainActor
 class BalanceBreakdownTests: XCTestCase {
-    @MainActor func testOnAppear() async throws {
+    func testOnAppear() async throws {
         let store = TestStore(
             initialState: .placeholder
         ) {
@@ -30,9 +32,6 @@ class BalanceBreakdownTests: XCTestCase {
         
         // expected side effects as a result of .onAppear registration
         await store.receive(.synchronizerStateChanged(.zero))
-        await store.receive(.updateLatestBlock) { state in
-            state.latestBlock = ""
-        }
 
         // long-living (cancelable) effects need to be properly canceled.
         // the .onDisappear action cancels the observer of the synchronizer status change.
@@ -41,35 +40,7 @@ class BalanceBreakdownTests: XCTestCase {
         await store.finish()
     }
 
-    @MainActor func testShieldFundsSucceed() async throws {
-        let store = TestStore(
-            initialState: .placeholder
-        ) {
-            BalanceBreakdownReducer(networkType: .testnet)
-        }
-
-        store.dependencies.sdkSynchronizer = .mock
-        store.dependencies.derivationTool = .liveValue
-        store.dependencies.mnemonic = .mock
-        store.dependencies.walletStorage.exportWallet = { .placeholder }
-        store.dependencies.walletStorage.areKeysPresent = { true }
-
-        await store.send(.shieldFunds) { state in
-            state.shieldingFunds = true
-        }
-        await store.receive(.shieldFundsSuccess) { state in
-            state.shieldingFunds = false
-            state.alert = AlertState.shieldFundsSuccess()
-        }
-        
-        // long-living (cancelable) effects need to be properly canceled.
-        // the .onDisappear action cancels the observer of the synchronizer status change.
-        await store.send(.onDisappear)
-        
-        await store.finish()
-    }
-
-    @MainActor func testShieldFundsFails() async throws {
+    func testShieldFundsFails() async throws {
         let store = TestStore(
             initialState: .placeholder
         ) {
@@ -83,10 +54,10 @@ class BalanceBreakdownTests: XCTestCase {
         store.dependencies.walletStorage.areKeysPresent = { true }
 
         await store.send(.shieldFunds) { state in
-            state.shieldingFunds = true
+            state.isShieldingFunds = true
         }
         await store.receive(.shieldFundsFailure(ZcashError.synchronizerNotPrepared)) { state in
-            state.shieldingFunds = false
+            state.isShieldingFunds = false
             state.alert = AlertState.shieldFundsFailure(ZcashError.synchronizerNotPrepared)
         }
 
@@ -97,25 +68,26 @@ class BalanceBreakdownTests: XCTestCase {
         await store.finish()
     }
 
-    @MainActor func testShieldFundsButtonDisabledWhenNoShieldableFunds() async throws {
+    func testShieldFundsButtonDisabledWhenNoShieldableFunds() async throws {
         let store = TestStore(
             initialState: .initial
         ) {
             BalanceBreakdownReducer(networkType: .testnet)
         }
 
-        XCTAssertFalse(store.state.shieldingFunds)
+        XCTAssertFalse(store.state.isShieldingFunds)
         XCTAssertFalse(store.state.isShieldableBalanceAvailable)
         XCTAssertTrue(store.state.isShieldingButtonDisabled)
     }
 
-    @MainActor func testShieldFundsButtonEnabledWhenShieldableFundsAvailable() async throws {
+    func testShieldFundsButtonEnabledWhenShieldableFundsAvailable() async throws {
         let store = TestStore(
             initialState: BalanceBreakdownReducer.State(
                 autoShieldingThreshold: Zatoshi(1_000_000),
-                latestBlock: L10n.General.unknown,
+                changePending: .zero,
+                isShieldingFunds: false,
+                pendingTransactions: .zero,
                 shieldedBalance: Balance.zero,
-                shieldingFunds: false,
                 synchronizerStatusSnapshot: .initial,
                 transparentBalance: Balance(
                     WalletBalance(
@@ -128,18 +100,19 @@ class BalanceBreakdownTests: XCTestCase {
             BalanceBreakdownReducer(networkType: .testnet)
         }
 
-        XCTAssertFalse(store.state.shieldingFunds)
+        XCTAssertFalse(store.state.isShieldingFunds)
         XCTAssertTrue(store.state.isShieldableBalanceAvailable)
         XCTAssertFalse(store.state.isShieldingButtonDisabled)
     }
 
-    @MainActor func testShieldFundsButtonDisabledWhenShieldableFundsAvailableAndShielding() async throws {
+    func testShieldFundsButtonDisabledWhenShieldableFundsAvailableAndShielding() async throws {
         let store = TestStore(
             initialState: BalanceBreakdownReducer.State(
                 autoShieldingThreshold: Zatoshi(1_000_000),
-                latestBlock: L10n.General.unknown,
+                changePending: .zero,
+                isShieldingFunds: true,
+                pendingTransactions: .zero,
                 shieldedBalance: Balance.zero,
-                shieldingFunds: true,
                 synchronizerStatusSnapshot: .initial,
                 transparentBalance: Balance(
                     WalletBalance(
@@ -152,8 +125,135 @@ class BalanceBreakdownTests: XCTestCase {
             BalanceBreakdownReducer(networkType: .testnet)
         }
 
-        XCTAssertTrue(store.state.shieldingFunds)
+        XCTAssertTrue(store.state.isShieldingFunds)
         XCTAssertTrue(store.state.isShieldableBalanceAvailable)
         XCTAssertTrue(store.state.isShieldingButtonDisabled)
+    }
+    
+    func testSyncingData() async throws {
+        let store = TestStore(
+            initialState: BalanceBreakdownReducer.State(
+                autoShieldingThreshold: Zatoshi(1_000_000),
+                changePending: .zero,
+                isShieldingFunds: true,
+                pendingTransactions: .zero,
+                shieldedBalance: Balance.zero,
+                synchronizerStatusSnapshot: .snapshotFor(state: .syncing(0.513)),
+                transparentBalance: Balance(
+                    WalletBalance(
+                        verified: Zatoshi(1_000_000),
+                        total: Zatoshi(1_000_000)
+                    )
+                )
+            )
+        ) {
+            BalanceBreakdownReducer(networkType: .testnet)
+        }
+
+        XCTAssertTrue(store.state.isSyncing)
+        XCTAssertEqual(store.state.syncingPercentage, 0.513 * 0.999)
+    }
+    
+    func testlastKnownSyncingPercentage_Zero() async throws {
+        let store = TestStore(
+            initialState: BalanceBreakdownReducer.State(
+                autoShieldingThreshold: Zatoshi(1_000_000),
+                changePending: .zero,
+                isShieldingFunds: true,
+                pendingTransactions: .zero,
+                shieldedBalance: Balance.zero,
+                synchronizerStatusSnapshot: .placeholder,
+                transparentBalance: Balance(
+                    WalletBalance(
+                        verified: Zatoshi(1_000_000),
+                        total: Zatoshi(1_000_000)
+                    )
+                )
+            )
+        ) {
+            BalanceBreakdownReducer(networkType: .testnet)
+        }
+
+        XCTAssertEqual(store.state.lastKnownSyncPercentage, 0)
+        XCTAssertEqual(store.state.syncingPercentage, 0)
+    }
+    
+    func testlastKnownSyncingPercentage_MoreThanZero() async throws {
+        let store = TestStore(
+            initialState: BalanceBreakdownReducer.State(
+                autoShieldingThreshold: Zatoshi(1_000_000),
+                changePending: .zero,
+                isShieldingFunds: true,
+                lastKnownSyncPercentage: 0.15,
+                pendingTransactions: .zero,
+                shieldedBalance: Balance.zero,
+                synchronizerStatusSnapshot: .placeholder,
+                transparentBalance: Balance(
+                    WalletBalance(
+                        verified: Zatoshi(1_000_000),
+                        total: Zatoshi(1_000_000)
+                    )
+                )
+            )
+        ) {
+            BalanceBreakdownReducer(networkType: .testnet)
+        }
+
+        XCTAssertEqual(store.state.lastKnownSyncPercentage, 0.15)
+        XCTAssertEqual(store.state.syncingPercentage, 0.15)
+    }
+    
+    func testlastKnownSyncingPercentage_FromSyncedState() async throws {
+        let store = TestStore(
+            initialState: BalanceBreakdownReducer.State(
+                autoShieldingThreshold: Zatoshi(1_000_000),
+                changePending: .zero,
+                isShieldingFunds: true,
+                lastKnownSyncPercentage: 0.15,
+                pendingTransactions: .zero,
+                shieldedBalance: Balance.zero,
+                synchronizerStatusSnapshot: .snapshotFor(state: .syncing(0.513)),
+                transparentBalance: .zero
+            )
+        ) {
+            BalanceBreakdownReducer(networkType: .testnet)
+        }
+
+        var syncState: SynchronizerState = .zero
+        syncState.syncStatus = .upToDate
+        let snapshot = SyncStatusSnapshot.snapshotFor(state: syncState.syncStatus)
+        
+        await store.send(.synchronizerStateChanged(syncState)) { state in
+            state.synchronizerStatusSnapshot = snapshot
+            state.syncStatusMessage = "Synced"
+            state.lastKnownSyncPercentage = 1.0
+        }
+    }
+    
+    func testlastKnownSyncingPercentage_FromSyncingState() async throws {
+        let store = TestStore(
+            initialState: BalanceBreakdownReducer.State(
+                autoShieldingThreshold: Zatoshi(1_000_000),
+                changePending: .zero,
+                isShieldingFunds: true,
+                lastKnownSyncPercentage: 0.15,
+                pendingTransactions: .zero,
+                shieldedBalance: Balance.zero,
+                synchronizerStatusSnapshot: .snapshotFor(state: .syncing(0.513)),
+                transparentBalance: .zero
+            )
+        ) {
+            BalanceBreakdownReducer(networkType: .testnet)
+        }
+
+        var syncState: SynchronizerState = .zero
+        syncState.syncStatus = .syncing(0.545)
+        let snapshot = SyncStatusSnapshot.snapshotFor(state: syncState.syncStatus)
+        
+        await store.send(.synchronizerStateChanged(syncState)) { state in
+            state.synchronizerStatusSnapshot = snapshot
+            state.syncStatusMessage = "Syncing"
+            state.lastKnownSyncPercentage = 0.545
+        }
     }
 }